+++ packages/annotation/src/arrows/index.tsx
@@ -1,2 +1,0 @@
-export * from './BlockArrow'
-export * from './types'