+++ packages/annotation/src/flowchart/index.tsx
@@ -1,2 +1,4 @@
+export * from './ArcArrow'
+export * from './BlockArrow'
 export * from './FlowPath'
 export * from './types'